+++ tradingstrategy/utils/summarydataframe.py
@@ -66,12 +66,17 @@ def create_summary_table(data: dict) -> pd.DataFrame:
     """Create a summary table from a human readable data.
 
     * Keys are human readable labels
+
     * Values are instances of :py:class:`Value`
+
+    TODO: We get column header "zero" that needs to be hidden.
     """
     formatted_data = {k: format_value(v) for k, v in data.items()}
     df = pd.DataFrame.from_dict(formatted_data, orient="index")
-    df.style.hide_index()
-    df.style.hide_columns()
+    # https://pandas.pydata.org/docs/dev/reference/api/pandas.io.formats.style.Styler.hide.html
+    df.style.hide(axis="index", names=True)
+    df.style.hide(axis="columns", names=True)
+    # df.style.hide_columns()
     df.style.set_table_styles([
         {'selector': 'thead', 'props': [('display', 'none')]}
     ])