+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add `get_single_pair_data` shortcut
 - Add `PandasPairUniverse.get_by_symbols` shortcut
 - Add `PandasPairUniverse.get_all_tokens` shortcut
+- Add `get_exchange_by_id` shortcut
 - Fix `filter_for_quote_tokens` to be more strict (was checking either side of the pair, not quote token)
 - API change `create_single_pair_universe` now to accept `pick_by_highest_vol`
 - API change: Rename `Universe.time_frame` -> `Universe.time_bucket` to be consistent